requirements.txt
torch>=1.9.0
transformers>=4.11.0
datasets>=1.11.0
tqdm>=4.62.0
pyyaml>=5.4.1
pytest>=7.0.0
pytest-cov>=4.0.0 